src/books_server/requirements.txt
python-multipart
fastapi
uvicorn
Pillow